requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.3
SQLAlchemy==2.0.30
alembic==1.13.2
python-multipart==0.0.9